requirements.txt
Flask==3.0.0
flask-cors==4.0.0
tensorflow-cpu==2.17.0
numpy==1.26.2
librosa==0.10.1
gunicorn==21.2.0
scikit-image